Simplify aggregated NEXRAD2 stats
See #3 for discussion.  The short story is that the origin host is now
ignored with the relay host used for both the origin and relay in the
case of NEXRAD2.  This cut the number of aggregated entries in the
hourly and daily tables by 75%.  It also greatly improves the plotting